Rename semerrgroup to parallel to better highlight its function
